+++ src/client/components/ah_accounts_select/ah_accounts_select.js
@@ -1,7 +1,7 @@
 /*
  * pwix:accounts-hub/src/client/components/ah_accounts_select/ah_accounts_select.js
  *
- * Select zero to n identities.
+ * Select zero to n accounts.
  * 
  * Parms:
  * - selectable: a list of user documents to be selected among
@@ -17,29 +17,34 @@
  */
 
 import _ from 'lodash';
-import 'multiple-select';
-import 'multiple-select/dist/multiple-select.min.css';
+import { multipleSelect } from 'multiple-select-vanilla';
 
+import { Logger } from 'meteor/pwix:logger';
 import { pwixI18n } from 'meteor/pwix:i18n';
 import { UIUtils } from 'meteor/pwix:ui-utils';
 
 import './ah_accounts_select.html';
 
+const logger = Logger.get();
+
 Template.ah_accounts_select.onCreated( function(){
     const self = this;
+    //logger.debug( self );
 
     self.AH = {
         $select: null,
+        ms: null,
+        selectables: new ReactiveVar( [] ),
         selectedIds: new ReactiveVar( [] ),
 
         // send the selection on each selection change
         // selected: an array of selected identities ids
         // items: an array of selected identities
-        triggerSelected( event, selected, dc ){
+        triggerSelected( event, selected ){
             let items = [];
             selected.forEach(( it ) => {
                 let found = false;
-                dc.selectable.every(( account ) => {
+                self.AH.selectables.get().every(( account ) => {
                     if( account._id === it ){
                         found = true;
                         items.push( account );
@@ -50,44 +55,43 @@ Template.ah_accounts_select.onCreated( function(){
             self.AH.$select.trigger( event, { selected: selected, items: items });
         }
     };
+
+    // track the selectables
+    self.autorun(() => {
+        self.AH.selectables.set( Template.currentData().selectable || [] );
+    });
 });
 
 Template.ah_accounts_select.onRendered( function(){
     const self = this;
 
-    self.AH.$select = self.$( '.ah-accounts-select select.multiple-select' );
-
-    // have the data context
-    let dc;
-    self.autorun(() => {
-        dc = Template.currentData();
-    });
-
     // prepare the multipleSelect configuration
+    const dataContext = Template.currentData();
     let conf = {
         selectAll: false,
         filter: true,
         classes: 'form-control',
-        maxHeight: 540,
+        autoAdjustDropHeight: true,
         onClick( data ){
-            self.AH.triggerSelected( 'ah-selected', self.AH.$select.multipleSelect( 'getSelects' ), dc );
+            const selected = self.AH.ms ? self.AH.ms.getSelects() : [];
+            self.AH.triggerSelected.bind( self )( 'ah-selected', selected );
         }
     };
-    if( Template.currentData().selectOptions ){
-        _.merge( conf, Template.currentData().options );
-    }
 
     // make sure that we have something before init the multipleSelect widget
     self.autorun(() => {
-        const selectable = Template.currentData().selectable;
-        if( selectable ){
-            UIUtils.DOM.waitFor( '.ah-accounts-select select.multiple-select' ).then(() => {
-                conf.placeholder = dc.select_ph || pwixI18n.label( I18N, 'dialogs.accounts_select_ph' );
-                self.AH.$select.multipleSelect( conf );
-            }).then(() => {
-                self.AH.$select.multipleSelect( 'open' );
-            });
+        const dataContext = Template.currentData();
+        if( dataContext.selectOptions ){
+            _.merge( conf, dataContext.options );
         }
+        UIUtils.DOM.waitFor( '.ah-accounts-select select.multiple-select' )
+            .then(() => {
+                self.AH.$select = self.$( '.ah-accounts-select select.multiple-select' );
+                conf.placeholder = dataContext.select_ph || pwixI18n.label( I18N, 'dialogs.accounts_select_ph' );
+                self.AH.ms = multipleSelect( self.AH.$select, conf );
+            }).then(() => {
+                self.AH.ms.open();
+            });
     });
 });
 
@@ -109,7 +113,7 @@ Template.ah_accounts_select.helpers({
 
     // return the list of selectable accounts
     itemsList(){
-        return this.selectable;
+        return Template.instance().AH.selectables.get();
     },
 
     // whether the current item is selected